Add external reference clock to ipbus_freq_ctr
Recent experience shows that the exact frequency of the IPBus clock
differs from board to board. This is no surprise, really, because
this clock tends to be derived from PCIe, AXI, etc. clocks which
are not necessarily very precise. It has a visible effect, though,
on the frequencies reported by ipbus_freq_ctr, which uses the IPBus
clock as reference.

The current modification adds a separate reference clock input to
ipbus_freq_ctr, so that a more precise clock can be used if so desired.