Fix the oscillator clock frequency in the VCU118 SGMII infrastructure
The frequency of the VCO clock frequency in the VCU118 SGMII infrastructure is incorrect. This
leads to complaints from Vivado about:
[DRC AVAL-168] MMCM_ADV Phase shift and divide attr checks: The MMCME4_ADV cell infra/clocks/mmcm has a fractional CLKFBOUT_MULT_F value
(3.333) which is not a multiple of the hardware granularity (0.125) and will be adjusted to the nearest supportable value. Please update the
design to use a valid value.

This patch fixes that by correcting the MMCME4 generic values provided in the infrastructure code.

The above can be easily reproduced (using Vivado 2021.2) by building the
top_vcu118_sgmii.dep example.