Support SYSMONs for all SLRs for UltraScale and UltraScale+
The previous version only instantiated a single SYSMON, even
for devices with multiple SLRs. The new implementation allows
the user to specify the number of SLRs to populate with SYSMON
instances using a VHDL generic.
The first SYSMON instance will be placed in the primary SLR,
and the other instances will cover the remaining SLRs in order.
The instance to address can be chosen with a control register
separate from the SYSMON DRP interface.

NOTE: This change is _not_ backwards compatible on purpose.
Having to add the generic to the instantiation makes the user of
the new (and recommended) functionality. The changes to the
address table also require some (minor) software changes.

NOTE: The restructuring of the address tables is a bit clumsy.
This seems to be the most straightforward appraoch for the multi-
architecture support without bending the available ipbb
functionality, though.