Add some 'missing' ipbb dependency files
This is not an attempt to provide ipbb dependency files for _all_ packages
and entities. It does, however, add several missing dependency files.

NOTE: The advantage of providing .dep files even for simple packages/entities
is that it makes it possible to later on update the implementation (and the
dependencies) without all downstream projects having to update their dependency
files.